+++ pkg/tlsutils/x509utils/X509CertificateHandler_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/asciich/asciichgolangpublic/logging"
 	"github.com/asciich/asciichgolangpublic/pkg/mustutils"
 	"github.com/asciich/asciichgolangpublic/testutils"
-	"github.com/asciich/asciichgolangpublic/tlsutils/x509utils"
+	"github.com/asciich/asciichgolangpublic/pkg/tlsutils/x509utils"
 )
 
 func getX509CertificateHandlerToTest(implementationName string) (handler x509utils.X509CertificateHandler) {
