+++ pkg/tlsutils/x509utils/X509Utils_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/asciich/asciichgolangpublic/pkg/cryptoutils"
 	"github.com/asciich/asciichgolangpublic/pkg/mustutils"
 	"github.com/asciich/asciichgolangpublic/testutils"
-	"github.com/asciich/asciichgolangpublic/tlsutils/x509utils"
+	"github.com/asciich/asciichgolangpublic/pkg/tlsutils/x509utils"
 )
 
 func Test_GetPublicKeyFromPrivateKey(t *testing.T) {
